[feat] UserService 추상화 및 TransactionUserService 기능 추가
